NftItem folder has been remade
