+++ client/pages/gallery.js
@@ -1,5 +1,5 @@
 import React from 'react'
-import NFTListContainer from '../components/NftPage/NftPage'
+import NFTListContainer from '../components/NftPage/NFTListContainer'
 
 function Gallery() {
   return <NFTListContainer />
